requirements.txt
elevenlabs
python-dotenv
paramiko
pyttsx3
